requirements.txt
python_dotenv==1.0.1
groq==0.13.0
